requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.17.0
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
newspaper3k==0.2.8
nltk==3.9.1
pillow==11.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
readability-lxml==0.8.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
